refactor again
Move and remove
