+++ MachineHypermediaToolkit/resource/HypermediaCollection.py
