+++ MachineHypermediaToolkit/resource/HypermediaRequest.py
