+++ MachineHypermediaToolkit/resource/HypermediaResource.py
