+++ MachineHypermediaToolkit/resource/PlainTextHandler.py
